// imxweb/projects/qbm/src/lib/authentication/auth-config-provider.interface.ts
/*
 * ONE IDENTITY LLC. PROPRIETARY INFORMATION
 *
 * This software is confidential.  One Identity, LLC. or one of its affiliates or
 * subsidiaries, has supplied this software to you under terms of a
 * license agreement, nondisclosure agreement or both.
 *
 * You may not copy, disclose, or use this software except in accordance with
 * those terms.
 *
 *
 * Copyright 2025 One Identity LLC.
 * ALL RIGHTS RESERVED.
 *
 * ONE IDENTITY LLC. MAKES NO REPRESENTATIONS OR
 * WARRANTIES ABOUT THE SUITABILITY OF THE SOFTWARE,
 * EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED
 * TO THE IMPLIED WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE, OR
 * NON-INFRINGEMENT.  ONE IDENTITY LLC. SHALL NOT BE
 * LIABLE FOR ANY DAMAGES SUFFERED BY LICENSEE
 * AS A RESULT OF USING, MODIFYING OR DISTRIBUTING
 * THIS SOFTWARE OR ITS DERIVATIVES.
 *
 */

import { AuthPropDataProvider } from '../session/auth-prop-data-provider.interface';
import { CustomAuthFlow } from './custom-auth-flow.interface';

export enum PreAuthStateType {
  PreAuth = 1,
  Auth = 2,
  Captcha = 3,
}
export interface AuthConfigProvider {
  name: string;
  display: string;
  isOAuth2?: boolean;
  customAuthFlow?: CustomAuthFlow;
  externalLogoutUrl?: string;
  authProps?: AuthPropDataProvider[];
  preAuthProps?: AuthPropDataProvider[];
  preAuthState?: PreAuthStateType;
}